Implemented Permissions
--Implemented Permissions

+++ RootsOfHealth/Scripts/ClientForms/ClientFormlist.js
@@ -28,30 +28,36 @@ function GetProgramTemplateList() {
                          <td width="15%">${item.ModifiedDate != null ? item.ModifiedDate.split("T")[0] : ""}</td>
                          <td width="10%">${(item.IsActive==1 && item.IsBaseTemplate == false) ? (item.Isactivated == 1 ? "Yes" : "No") : ""}</td>
                             <td width="30%"><div>`;
-                    programs += `<a href="javascript:void(0)" onclick="ViewProgramContent(${item.TemplateID},\'${(item.FormName == null && item.IsBaseTemplate == true) ? "Base Template" : item.FormName}'\)" class="btn btn-success text-white" style="cursor:pointer;">VIEW</a>`
-                    if (item.IsBaseTemplate) {
+                    if (canViewClientForm == 'True')
+                    {
+                        programs += `<a href="javascript:void(0)" onclick="ViewProgramContent(${item.TemplateID},\'${(item.FormName == null && item.IsBaseTemplate == true) ? "Base Template" : item.FormName}'\)" class="btn btn-success text-white" style="cursor:pointer;">VIEW</a>`
+                    }
+                    if (item.IsBaseTemplate && canEditClientForm == 'True') {
                         ;
                         item.IsActive = item.IsActive == null ? false : true;
                         programs += `<a href="/Client/ClientFormBaseTemplate?templateid=${item.TemplateID}&Status=${item.IsActive}"  class="btn btn-success text-white" style="cursor:pointer;">MODIFY</a>`
 
-                    } else {                        
-                        programs += `<a href="javascript:void(0)" onclick="ProceedProgram({TemplateID:${item.TemplateID},TemplateTable:\'${item.TemplateTable}\',FormName:\'${item.FormName}\',ClientFormID:${item.ClientFormID},IsBaseTemplate:${item.IsBaseTemplate}})"  class="btn btn-success text-white" style="cursor: pointer; ${item.Isactivated == true ? "display:none;" : ""} ">MODIFY</a>`
+                    } else {
+                        if (canEditClientForm == 'True')
+                        {
+                            programs += `<a href="javascript:void(0)" onclick="ProceedProgram({TemplateID:${item.TemplateID},TemplateTable:\'${item.TemplateTable}\',FormName:\'${item.FormName}\',ClientFormID:${item.ClientFormID},IsBaseTemplate:${item.IsBaseTemplate}})"  class="btn btn-success text-white" style="cursor: pointer; ${item.Isactivated == true ? "display:none;" : ""} ">MODIFY</a>`
+                        }
                     }
                     
-                    if (item.IsActive == 1 && item.IsBaseTemplate == false) {
+                    if (item.IsActive == 1 && item.IsBaseTemplate == false && canEditClientForm == 'True') {
                         if (item.Isactivated == true) {
                             programs += `<a href="javascript:void(0)"  onclick="SetTemplateStatus(${item.TemplateID},${false})"  class="btn btn-success text-white" style="cursor:pointer;">DEACTIVATE</a>`;
                         } else if (item.Isactivated == 0) {
                             programs += `<a href="javascript:void(0)" onclick="SetTemplateStatus(${item.TemplateID},${true})"  class="btn btn-success text-white" style="cursor:pointer;">ACTIVATE</a>`;
                         }
-                    } else if (item.IsActive == 0 && item.IsBaseTemplate == false){
+                    } else if (item.IsActive == 0 && item.IsBaseTemplate == false && canEditClientForm == 'True'){
                         programs += `<a href="javascript:void(0)" onclick="alertInprogressStatus()"  class="btn btn-success text-white" style="cursor:pointer;">ACTIVATE</a>`;
                     }
 
-                    if (!item.IsBaseTemplate) {
+                    if (!item.IsBaseTemplate && canDeleteClientForm == 'True') {
                         programs += `<a href="javascript:void(0)" onclick="DeleteProgram(${item.ClientFormID},this)"  class="btn btn-success text-white" style="cursor:pointer;">Delete</a>`;
                     }
-                    if (!item.IsBaseTemplate && item.IsActive == 1) {
+                    if (!item.IsBaseTemplate && item.IsActive == 1 && canEditClientForm == 'True') {
                         programs += `<a href="javascript:void(0)" onclick="SetStatus(${item.ClientFormID})"  class="btn btn-success text-white" style="cursor:pointer;">Manage Status</a>`;
                     }
